requirements.txt
discord.py>=2.3.0,<3
certifi>=2024.0.0,<2026
httpx>=0.27.0,<1
python-dotenv>=1.0.0,<2
pytest>=8.0.0,<9
